Canonicalise broadcom DRMbuf format modifiers
When checking if a DMAbuf's format modifiers are compatible with the
output, apply canonicalisation to Broadcom modifiers.  These modifiers
include stripe height as a parameter and the outputs support all
parameter values without explicitly listing them.

This enables direct scanout of DMAbufs from the HEVC decoder on the
Raspberry Pi 4.