Allow scaling of scanout buffers
Set the CRTC_{W,H} parameters on the layers to be scanned out to the
output size, not the buffer size, so KMS will scale them if necessary.
Then remove the check which prevents direct scanout when the buffer is
not the same size as the output.

Tested this with direct scanout of hardware-decoded h.264 and HEVC on
Raspberry Pi 4.